requirements.txt
fastapi
uvicorn
python-multipart
clarifai
python-dotenv